backend/requirements.txt
# =============================================================================
# ShopSense AI - Backend Dependencies
# =============================================================================
# Core Framework
# =============================================================================
Flask==3.1.0
Flask-CORS==5.0.0
Flask-SocketIO==5.4.1
Werkzeug==3.1.3

# =============================================================================
# Authentication & Security
# =============================================================================
PyJWT==2.10.1
bcrypt==4.3.0
Flask-Limiter==3.7.0
Flask-Talisman==1.1.0
email-validator==2.2.0

# =============================================================================
# Database
# =============================================================================
pymongo==4.10.1

# =============================================================================
# Data Processing & Analytics
# =============================================================================
pandas==2.2.3
numpy==2.2.3
plotly==5.24.1
prophet==1.1.5
cmdstanpy==1.2.5

# =============================================================================
# AI & Machine Learning
# =============================================================================
google-generativeai==0.8.4
protobuf==5.29.3

# =============================================================================
# Utilities
# =============================================================================
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
eventlet==0.40.4

# =============================================================================
# Export & Reporting
# =============================================================================
openpyxl==3.1.5

# =============================================================================
# Validation
# =============================================================================
pydantic==2.10.6
pydantic-settings==2.7.1

# =============================================================================
# Testing (Development)
# =============================================================================
pytest==8.3.4
pytest-cov==6.0.0
pytest-flask==1.3.0
responses==0.25.7

# =============================================================================
# Development Tools
# =============================================================================
black==24.10.0
flake8==7.1.1
mypy==1.14.1
